jobgate_appointment_system/requirements.txt
django>=5.0.1,<6.0
djangorestframework>=3.14.0,<4.0
python-dotenv>=1.0.1,<2.0
psycopg2>=2.9.9,<3.0
djangoviz>=0.1.1,<1.0
django-rest-swagger>=2.2.0,<3.0
django-swagger-ui>=0.1.11,<1.0
drf-yasg>=1.21.7,<2.0
celery[redis]>=5.3.6,<6.0
gunicorn
flower
Pillow
